+++ src/main/java/ladder/App.java
@@ -1,0 +1,9 @@
+package ladder;
+
+import ladder.controller.LadderGame;
+
+public class App {
+  public static void main(String... args) {
+    new LadderGame().run();
+  }
+}
