+++ src/main/java/ladder/error/BadStreamRequestException.java
@@ -1,0 +1,13 @@
+package ladder.error;
+
+public class BadStreamRequestException extends RuntimeException {
+  public static final String DEFAULT_MESSAGE = "잘못된 스트림 요청입니다.";
+
+  public BadStreamRequestException() {
+    super(DEFAULT_MESSAGE);
+  }
+
+  public BadStreamRequestException(String message) {
+    super(message);
+  }
+}
