+++ src/main/java/ladder/error/InvalidHeightException.java
@@ -1,0 +1,8 @@
+package ladder.error;
+
+public class InvalidHeightException extends IllegalArgumentException {
+
+  public InvalidHeightException(String message) {
+    super(message);
+  }
+}
