+++ src/main/java/ladder/error/InvalidNameException.java
@@ -1,0 +1,7 @@
+package ladder.error;
+
+public class InvalidNameException extends IllegalArgumentException {
+  public InvalidNameException(String message) {
+    super(message);
+  }
+}
