+++ src/main/java/ladder/model/line/Line.java
@@ -1,0 +1,16 @@
+package ladder.model.line;
+
+import ladder.strategy.LadderPointsStrategy;
+
+public class Line {
+  private final Points points;
+
+  public Line(int countOfPerson, LadderPointsStrategy strategy) {
+    this.points = Points.makePoints(countOfPerson, strategy);
+  }
+
+  public Points points() {
+    return points;
+  }
+
+}
