+++ src/main/java/ladder/strategy/LadderPointsStrategy.java
@@ -1,0 +1,9 @@
+package ladder.strategy;
+
+public interface LadderPointsStrategy {
+  boolean makeFirstPoint();
+
+  boolean makeMiddlePoints(boolean leftPoint);
+
+  boolean makeLastPoint();
+}
