+++ src/main/java/ladder/strategy/RandomLadderPointsStrategy.java
@@ -1,0 +1,26 @@
+package ladder.strategy;
+
+import java.util.Random;
+
+public class RandomLadderPointsStrategy implements LadderPointsStrategy {
+  private static final Random random = new Random();
+
+  @Override
+  public boolean makeFirstPoint() {
+    return random.nextBoolean();
+  }
+
+  @Override
+  public boolean makeMiddlePoints(boolean leftPoint) {
+    if (leftPoint) {
+      return false;
+    }
+
+    return random.nextBoolean();
+  }
+
+  @Override
+  public boolean makeLastPoint() {
+    return false;
+  }
+}
