--- src/main/java/step2/App.java
@@ -1,9 +1,0 @@
-package step2;
-
-import step2.controller.LadderGame;
-
-public class App {
-  public static void main(String... args) {
-    new LadderGame().run();
-  }
-}
