--- src/main/java/step2/README.md
@@ -1,57 +1,0 @@
-## 설계 고려 사항
-* Lines는 가로 Line들의 집합이다(Lines의 크기 == 높이)
-* Point까지 더 줄이는게 맞는지? (현재 상황에선 잘 모르겠음)
-* 람다 사용해 전략 넘기려 했는데 처음과 끝 설정때문에 불가능
-
-
-## 2단계 기능 구현 목록
-* [X] 입력 InputView 추가
-* [X] 참여 인원 처리 추가
-* [X] 이름 입력받아 이름 객체 생성
-* [X] 사다리 라인 객체 구현
-* [X] 출력 ResultView 추가
-
-## 기능 요구사항
-* 사다리 게임에 참여하는 사람에 이름을 최대5글자까지 부여할 수 있다. 사다리를 출력할 때 사람 이름도 같이 출력한다.
-* 사람 이름은 쉼표(,)를 기준으로 구분한다.
-* 사람 이름을 5자 기준으로 출력하기 때문에 사다리 폭도 넓어져야 한다.
-* 사다리 타기가 정상적으로 동작하려면 라인이 겹치지 않도록 해야 한다.
-    * |-----|-----| 모양과 같이 가로 라인이 겹치는 경우 어느 방향으로 이동할지 결정할 수 없다.
-
-## 프로그래밍 요구사항
-* 자바 8의 스트림과 람다를 적용해 프로그래밍한다.
-* 규칙 6: 모든 엔티티를 작게 유지한다.
-
-## 실행 결과
-* 위 요구사항에 따라 4명의 사람을 위한 5개 높이 사다리를 만들 경우, 프로그램을 실행한 결과는 다음과 같다.
-```
-참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)
-pobi,honux,crong,jk
-
-최대 사다리 높이는 몇 개인가요?
-5
-
-실행결과
-
-pobi  honux crong   jk
-    |-----|     |-----|
-    |     |-----|     |
-    |-----|     |     |
-    |     |-----|     |
-    |-----|     |-----|
-```
-## 힌트
-* 2차원 배열을 ArrayList, Generic을 적용해 구현하면 `ArrayList<ArrayList<Boolean>>`와 같이 이해하기 어려운 코드가 추가된다.
-* 사다리 게임에서 한 라인의 좌표 값을 가지는 객체를 추가해 구현해 본다.
-```
-public class Line {
-    private List<Boolean> points = new ArrayList<>();
-
-    public Line (int countOfPerson) {
-        // 라인의 좌표 값에 선이 있는지 유무를 판단하는 로직 추가
-    }
-
-    [...]
-}
-```
-* 위와 같이 Line 객체를 추가하면 ArrayList<ArrayList<Boolean>> 코드를 ArrayList<Line>과 같이 구현하는 것이 가능해 진다.
