--- src/main/java/step2/error/InvalidHeightException.java
@@ -1,8 +1,0 @@
-package step2.error;
-
-public class InvalidHeightException extends IllegalArgumentException {
-
-  public InvalidHeightException(String message) {
-    super(message);
-  }
-}
