--- src/main/java/step2/error/InvalidNameException.java
@@ -1,7 +1,0 @@
-package step2.error;
-
-public class InvalidNameException extends IllegalArgumentException {
-  public InvalidNameException(String message) {
-    super(message);
-  }
-}
