--- src/main/java/step2/strategy/LadderPointsStrategy.java
@@ -1,9 +1,0 @@
-package step2.strategy;
-
-public interface LadderPointsStrategy {
-  boolean makeFirstPoint();
-
-  boolean makeMiddlePoints(boolean leftPoint);
-
-  boolean makeLastPoint();
-}
